+++ downloader/sysproc_windows.go
@@ -1,0 +1,13 @@
+//go:build windows
+
+package downloader
+
+import (
+	"os/exec"
+	"syscall"
+)
+
+// setHideWindow 在 Windows 上隐藏 CMD 窗口
+func setHideWindow(cmd *exec.Cmd) {
+	cmd.SysProcAttr = &syscall.SysProcAttr{HideWindow: true}
+}